+++ src/enabled.rs
@@ -43,6 +43,9 @@ mod tlskdf;
 #[cfg(feature = "mlkem")]
 mod mlkem;
 
+#[cfg(feature = "mldsa")]
+mod mldsa;
+
 use mechanism::Mechanisms;
 use object::ObjectFactories;
 
@@ -90,6 +93,9 @@ fn register_all(mechs: &mut Mechanisms, ot: &mut ObjectFactories) {
 
     #[cfg(feature = "mlkem")]
     mlkem::register(mechs, ot);
+
+    #[cfg(feature = "mldsa")]
+    mldsa::register(mechs, ot);
 
     #[cfg(feature = "fips")]
     fips::register(mechs, ot);